fetching data on submit buttin
